requirements.txt
Django==5.0.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
pytils==0.4.1
django-phonenumber-field==8.0.0
phonenumberslite==8.13.42
Pillow==10.4.0